requirements.txt
tf-nightly==2.7.0.dev20210801
keras-nightly==2.7.0.dev2021081000      
scikit-learn
matplotlib==3.3.4            
numpy==1.19.2           
h5py==3.1.0